Adding inglourious and basterds to the dictionary.  Adding scoring Levenshtein algorithm.  Matching from file path to TMDB and back.

+++ src/shared/MatchScore.cs
@@ -7,7 +7,7 @@ namespace shared
 {
     public class MatchScore<ContainedType> where ContainedType : class
     {
-        public long Hits { get; set; } = 0;
+        public double Hits { get; set; } = 0;
         public ContainedType? Details { get; set; }
     }
 }
